RTS-16532 Uib_study nodes should not get ou images in related content
For the areas foremployees and education we use an image from the parent
ou of a node when shown as related content in a node. Uib_study nodes
belong to the education area, but related content nodes should use their
own image.

This commit also includes styling done in this issue in SG
(commit:531f49e).