RTS-16978 Make sure study facts with no value is really hidden.
Currently when a study fact does not have a value for the current
semester but a variant exist for previous or next semester the study
fact shows as a heading with no corresponding value. In this case the
heading should also be gone. This commit makes the test for these cases
more robust for the obvious fields.