RTS-16919 Use correct info types for bilateral agreements
The following info types are the only that should be used for bilateral
agreements:

- FAVT_HIST
- FAVT_MOB
- FAVT_TYPE

And the same info types should never be rendered by any other study
nodes of type exchange. This is solved by only keeping the wanted type
for the bilateral agreements and adding the specific types to the
splice_list for the other exchange nodes to be sure the info types are
not rendered for these study categories.