RTS-17060 Made it possible to use node id in autocomplete
It may be difficult to get relevant hits for editors in autocomplete
fields when content titles are to generic. Makeing the node id
searchable will improve this.

This is solved by adding nid as a field in the relevant view instances
(hidden from display) and and adding it to search fields in the entity
reference.