RTS-16964 Made sure contruction table in study programmes only show once
We have two possible tabs for a contruction table to be displayed in a
study programme. Due to a too general if statement the contruction table
was displayed in both tabs. By adding a variable for the tab we are sure
the table only get displayed at the preferred location.

We also move the contruction table to be displayed after any video and
testimonial present on the study programme.