RTS-17345 Emergency tweaking
Date as header changed from h2 to h4 in order to make the date
less dominating in the messages on external pages.

The "read more" link for the english version was changed in order
to direct the link to the english article.

The archive link was changed to absolute = TRUE. From API:
The absolute property defaults to false. Whether to force the
output to be an absolute link (beginning with http). Useful
for links that will be displayed outside the site, such as in
an RSS feed.