RTS-17360 Update to duration value for a study programme
We had an error in the formula calculating the duration of a study
programme which was affecting parttime study programmes. Also we would
like to get the value from the field 'varighet' in fs-pres if the value
is set. If not the output should be generated based on the formula.