RTS-15411 Added settings for manualcrop
It is only possible to use one manual crop for one image style. We want
the image to be cropped the same way in every image style except
wide_thumbnail. This should be cropped, but also use the focal point to
get the correct style. These settings reflect this wish.