RTS-17971 Prevent users from deleting file that are in use
We have an issue with users deleting files that are in use despite the
warning telling them it may cause problems. To remedy this situation we
remove the confirm button from the delete file form until the file is no
longer in use in any content.

We need to investigate why the file_managed and the file_usage tables are
not updated correctly. This fix makes it possible to still stop people
deleting files that are in use.