RTS-18080 Image text should not display for view mode full_width_image
We hide image text for view modes used in the area pages. We need to do
this for the full_width_image view mode as well as the copyright is now
a required field and although we want the information we do not want to
display the information in the area page.