RTS-18141 Make sure image caption is show beside the image
After we added image caption for all media types the new approach made
the main media display with the media caption below the image and not
beside the image. This was due to a minor error in the logic in
preprocess_field. The test was designed to consider the initial values
in field_uib_description and did not consider the values after we
inserted a possible image caption to the field.

To make sure the correct class is inserted we need to test the values in
the items array, not the elements array.