RTS-18178 New view mode quote used in testimonials
As we need to make testimonials look good in promoted content we need to
create a new view mode to use when testimonials are displayed in their
current teaser form. We call this new view mode quote.

Setting up this view mode involves creating the view mode and exporting
the settings for the view mode in features, configuring the view mode
for the content type and exporting the settings in features, and using
the new view mode when displaying a testimonial.