RTS-19025 Updates to feature article code
To accommodate the demand that only feature areas and the news area
should be able to use the new article we needed to change how to
identify a feature article. This is now done by selecting article type
news and add the area. If the area is a feature area the checkbox
Feature article appears.

This commit also includes:
- Minor tweaks to markup in editor templates
- Minor tweaks to settings for feature article
- Minor tweaks to theme
- commit:2d0de0f from SG