RTS-19367 Option feature article should be determined by area type
As not all feature areas wanted to change area when the option was
available we needed to hard code what areas should be able to use this
option. Now all areas have changed type and we should determine the
availability of the feature article option based on area type. An option
to include the news areas in the resulting array is included as a
parameter to the function.