RTS-19426 Better field restrictions in area edits
The control structure used to hide and show fields for different areas
are fragile. By replacing the switch structure with simple if
statements it should be easier to maintain.

This also makes sure that the promoted articles field only is available
for areas of type newspage or frontpage and the nids identifying the
research areas.

It also changes the autocomplete source of the promoted article field.