RTS-19492 Option for disclaimer text in exam info
SA wants to add a disclaimer text to the exam information displayed in
courses. To make it possible to alter or remove this text on the fly the
feature is implemented as two variables. One for the norwegian string
(uib_exam_disclaimer) and one for the english (uib_exam_disclaimer:en).
If either of these does not exist or is set to FALSE no string will
show.