RTS-19053 Admin needs access to search hits
When we added the field for search hits the field was hidden for all
users. This included administrator. For all jobs that required access by
admin on node types with this field the jobs did not execute (ex.
scheduling on nodes). By granting admin access to this field the problem
should disappear.