RTS-19453 Update to exam information list
We want the same exam information to be display for each course in the
filter list as on the course page. To achieve this we use the same
function to generate the data both places. To be able to hide some
information in the list view a flag is added to indicate a list view.

includes commit:bbef2b2 from SG