RTS-19655 Added language option to tp url
We have added a variable to update the tp url on the fly. But tp uses a
different value for norwegian (no) than w3 (nb) so to be able to get the
correct schedule link we have to add a test to add the correct value for
the language parameter in the url.

Also added is a drush command in post-upgrade to delete the current tp
variable. We will the use the updated default value.