RTS-19701 Some exams does not appear in the generated exam lists
Any exams containing a date in either the root or in a exam part should
be displayed in the generated lists. The order of conditions for showing
the exams made exams with date only in the root and not in any exam part
disappear from the list. By reversing the order of conditions we make
sure all exams appear in the lists.