RTS-19775 Calendar block should only execute in context of an area
The test to done to decide if the calendar block should execute or not
would always evaluate to true if the current node belonged to an area
and affecting events displayed as a full page.

By supplying an additional support function to this test we get the
correct output and full page events are no longer affected by the
calendar block.