RTS-19867 Added admin view for research field taxonomy
To save time in finding terms for users that need to change something in
registered terms we want to add this simple tool. By implementing a view
with an exposed filter for the term name a level 1 user should be able
to search for a term and not click through an unknown number of pages to
find the correct term.