RTS-20166 New field and taxonomy for node type study
To implement the new grouping based on different categories than the NUS
categories used today we need a new field (field_uib_study_tag). The possible
content of this field should be based on categories maintained in a new
taxonomy (uib_study_categories). The taxonomy terms have one
translatable field to use as group heading in the filter page.