RTS-20290 Minor markup changes to study listing
This commit implements the wanted to the top link for the alphabetical
list of study programmes. The scroll function is implemented as an
anchor link and the javascript adds a animated scroll to the click
function. The link should be functional when javascript is turned off.

Some minor tweaks to the markup to facilitate some final style issues in
the grouped listing.

Inlucdes SG commit:50a0cf7