RTS-20296 Added class to footer-wrapper
Added the class in order for the dropdown menu to work on mobile
without adding a separate script to solve this.

Altered mobile_menu.js to only stop default behaviour when the
window size is less than 600.

This will stop the collapsing of menu on desktop and will keep
top menu items as links

prevent anchor links in menu to jump to top

Includes SG commit:35937f2