RTS-20295 Bug in generating urls and redirects
The urls for study specializations are generated on the form of
studyprogramme/[main study code]/[specialization study code]

The study listing pages and the new aliases and redirects for the
study specializations did not consider this pattern. By adding path to
the json api for the study listings we can get the correct url when the
specializations are shown and we may generate new aliases and redirects
based on the old ones.