RTS-20338 Show correct study titles
Due to a bug in our import of study nodes study nodes where the title is
the same in norwegian and english do not get a title in the english
version. On the initial import no language specific title is set, only
the und version of the field.

This is a workaround until we solve the issue setting correct language
versions in the import.