RTS-20107 KMD nodes should get KMD url
When we import the KMD articles and events we also update the elastic
index. The url saved to the elastic index is the url to the stub of the
article in w3. We need to use the href saved in the field_uib_kmd_data
where this field contains any data.