RTS-20365 Page to review outdated content
We have several jobs that touch a node and creates a revision of the
page. The author of the revision is registered as Administrator or
Anonymous. This makes it difficult to find nodes that has not been
updated on a regular basis.

The solution to this problem is to create a page that outputs nodes with
more than six months since the last revision not done by the system.