RTS-14344 Change page title bar title to reflect area
Change title bar title on pages where the page is a sub-page to
an area (i.e. area/calendar, area/news-archive, area/courses)

Added in preprocess_html, but the work is done in a function in
order to limit the line-size of the preprocess_html function

Alter the function uib_w3__get_theme_suggested_title to replace
the duplicated code to replace title.

Adding a parameter $with_name in the function call, setting this
default to true, to get "| University of Bergen" at the end of the
title in the page title bar.

Stop theme suggested title from rendering on area