RTS-20621 Added a tab with outdated content to areas
A while ago we added a page listing outdated nodes to benefit the users
with the role 'superbruker'. Now we want to a tab to each area so that
users with access to edit that area can get an overview of outdated
nodes in their area. This is achieved by a slight rewrite of the
existing code.