RTS-20867 Fix for bug in outdated pane
When we introduced the outdated pane we also introduced a bug that made
nodes with any revision done by a real user for more than six months ago
appear (always). We only want nodes with the newest revision done by a
real user for more than six months ago to appear. This tweak to the
query should fix that.