RTS-20846 Small media as teaser
When there is no main media in articles, and the article only uses
small images. When the article is used as a profiled article, one
of the smaller images are used as main media.

The main media markup is wrapped in a HTML a-element in order to
link this image to the main article