RTS-21404 Better media handling for media in text
We want simplify the use of media captions when the image is inserted
into the text. The use of templates proved to be a steep learning curve
and now that this feature is going to be available to more content
managers we needed to rethink the approach. The solution was to add text
fields as we use with regular media in all content. This involves:

- remove the templates in wysiwyg
- add a new field with the media captions
- update the way captions is inserted into the text
- minor tweaks to styling

Includes commit:ef61d7a in SG