RTS-21389 Study facts for studyprogrammes
The new layout of study programmes requires two blocks of infomation. A
study facts line in the top of the programme and a block that may be
inserted in the text provided by a template in the editor. The data is
provided by fs-pres and new fields in w3. This issue implements:

- the new fields required in w3
- update to the study facts line
- expand the template string in editor to markup on view

Includes commit:bd3d1e3 in SG