RTS-20722 Change employee pages visual RTS-20722 New regions for employee pages with templates RTS-20722 Changed header title for employee pages
Squashing commits

Created a page template page__employee.tpl.php
Changed in template.php to use the page template when area is
employee pages

Added new regions in the theme.info (uib_w3.info) in order to have
a more separate rendering capability for the employee pages

Changed the template to render the new regions.
Region employee header in stead of region header
Region employee subheader in sted of subheader

Added class header__wrapper__employee to in the page__employee
template in order to reuse existing css and add specific css
for employee pages

Reused as much as I could of the code required to add items to
regions, but changed the content where it was required.

Includes commit:9c92f63 from SG