RTS-21505 Fixed access control for delete pane in node type uib_article
The access control for the delete pane in uib_article has probably never
been functional. Only admin and the role superbruker would have been
able to access the pane. This commit intends to grant access to the pane
for user 1, users with the superbruker role and the role
innholdsprodusent in the owner area of the node. Authenticated and
anonymous users should not be granted access to the pane.