RTS-20574 Add masemap to user profiles
Adding mazemap link field to the user profile

In the template.php there is code extracting the POI from the user
input link for the mazemap. If the link is a mazemap, the POI is
extracted directly from the URL provided

If the link provided is with the "share link" option. The user will
provide a bit.ly url. If this is the case, the forwarding link is
opened and checked if the page content on the forward page contains
both the user.mazemap.com link and a POI