RTS-21847 Fix for study codes in study text
As we use field translation for study nodes the text filters is run in
context of the current language every time a study node is saved. This
causes the study code filter to expand to the path of a course in the
current language every time a study programme is saved. Either english
or norwegian version is always wrong.