RTS-22190 Remove fields from studyprogrammes
As we have introduced a new template for study programmes we need to
remove some fields from view and edit modes for a study programme that
use the new template. Most of the fields were already removed in view
mode, but this aims to remove the remaining in view mode and all the
fields in edit mode for programmes checked to use the new template.