Adding function to get breadcrumb menu
If the last crumb in the breadcrumb is linking to the current
path, but if not from the current menu tree, the menu is checked
to see if it contains an item to current path, and then added
if it exist

If the item path is (node/%/persons) the menu item with the link
is added, if it's there.

The function __uib_get_current_menu_crumb takes no parameters,
but loads current area, and current path. Then checks if there
is a menu item linking to the current path.