RTS-22409 Fix for javascript failure in image_display_formats.js
The javascript introduced in commit:14f28e0b was not contained to the
correct form. It was attached to any form causing javascript failures in
other scripts as well. The most critical failure was the chatbot hubro.

By splitting the javascript functionality into two scripts and attaching
the functions to the correct forms the error was fixed.