RTS-22658 Move and delete not translated data in study programme field
At first the field field_uib_study_language was not translateable. This
was later altered but the data already inserted was not handled. We need
to handle this now as changes made to the study language mat not be
reflected as the old data may leak into view.

This commit adds a script to move data from the not translated version
of the field to the published language versions of the field.