Upgraded bower version
A node.js test version of the jso library provided by dataporten contains a
reference to an old version of bower. We have never used this version
and no other node modules referenced in package.json  or any other libraries
referenced in bower. This is altered because github is issuing warnings
about the bower version.